fix: handle apps with empty update logs
